Ported API doc to Vue.js (basic)
